services/api_github/requirements.txt
fastapi
uvicorn
PyGithub
python-dotenv
